# Experimenting with spans and StringSegment