--- a/src/Esprima/Ast/Directive.cs
+++ b/src/Esprima/Ast/Directive.cs
@@ -1,15 +1,16 @@
 ﻿using System.Runtime.CompilerServices;
+using Microsoft.Extensions.Primitives;
 
 namespace Esprima.Ast;
 
 public sealed class Directive : ExpressionStatement
 {
-    public Directive(Expression expression, string value) : base(expression)
+    public Directive(Expression expression, StringSegment? value) : base(expression)
     {
         Value = value;
     }
 
-    public string Value { [MethodImpl(MethodImplOptions.AggressiveInlining)] get; }
+    public StringSegment? Value { [MethodImpl(MethodImplOptions.AggressiveInlining)] get; }
 
     protected override ExpressionStatement Rewrite(Expression expression)
     {