--- a/src/Esprima/Ast/Jsx/JsxText.cs
+++ b/src/Esprima/Ast/Jsx/JsxText.cs
@@ -1,20 +1,21 @@
 ﻿using System.Diagnostics;
 using System.Runtime.CompilerServices;
 using Esprima.Utils.Jsx;
+using Microsoft.Extensions.Primitives;
 
 namespace Esprima.Ast.Jsx;
 
 [DebuggerDisplay("{Raw,nq}")]
 public sealed class JsxText : JsxExpression
 {
-    public JsxText(string? value, string raw) : base(JsxNodeType.Text)
+    public JsxText(StringSegment? value, StringSegment raw) : base(JsxNodeType.Text)
     {
         Value = value;
         Raw = raw;
     }
 
-    public string? Value { [MethodImpl(MethodImplOptions.AggressiveInlining)] get; }
-    public string Raw { [MethodImpl(MethodImplOptions.AggressiveInlining)] get; }
+    public StringSegment? Value { [MethodImpl(MethodImplOptions.AggressiveInlining)] get; }
+    public StringSegment Raw { [MethodImpl(MethodImplOptions.AggressiveInlining)] get; }
 
     internal override Node? NextChildNode(ref ChildNodes.Enumerator enumerator) => null;