--- a/src/Esprima/Ast/Literal.cs
+++ b/src/Esprima/Ast/Literal.cs
@@ -2,6 +2,7 @@
 using System.Runtime.CompilerServices;
 using System.Text.RegularExpressions;
 using Esprima.Utils;
+using Microsoft.Extensions.Primitives;
 
 namespace Esprima.Ast;
 
@@ -10,34 +11,34 @@ public sealed class Literal : Expression
     private static readonly object s_boxedTrue = true;
     private static readonly object s_boxedFalse = false;
 
-    internal Literal(TokenType tokenType, object? value, string raw) : base(Nodes.Literal)
+    internal Literal(TokenType tokenType, object? value, StringSegment raw) : base(Nodes.Literal)
     {
         TokenType = tokenType;
         Value = value;
         Raw = raw;
     }
 
-    public Literal(string value, string raw) : this(TokenType.StringLiteral, value, raw)
+    public Literal(StringSegment value, StringSegment raw) : this(TokenType.StringLiteral, value, raw)
     {
     }
 
-    public Literal(bool value, string raw) : this(TokenType.BooleanLiteral, value ? s_boxedTrue : s_boxedFalse, raw)
+    public Literal(bool value, StringSegment raw) : this(TokenType.BooleanLiteral, value ? s_boxedTrue : s_boxedFalse, raw)
     {
     }
 
-    public Literal(double value, string raw) : this(TokenType.NumericLiteral, value, raw)
+    public Literal(double value, StringSegment raw) : this(TokenType.NumericLiteral, value, raw)
     {
     }
 
-    public Literal(BigInteger value, string raw) : this(TokenType.BigIntLiteral, value, raw)
+    public Literal(BigInteger value, StringSegment raw) : this(TokenType.BigIntLiteral, value, raw)
     {
     }
 
-    public Literal(string raw) : this(TokenType.NullLiteral, null, raw)
+    public Literal(StringSegment raw) : this(TokenType.NullLiteral, null, raw)
     {
     }
 
-    public Literal(string pattern, string flags, object? value, string raw) : this(TokenType.RegularExpression, value, raw)
+    public Literal(StringSegment pattern, StringSegment flags, object? value, StringSegment raw) : this(TokenType.RegularExpression, value, raw)
     {
         // value is null if a Regex object couldn't be created out of the pattern or options
         Regex = new RegexValue(pattern, flags);
@@ -45,10 +46,10 @@ public sealed class Literal : Expression
 
     public TokenType TokenType { [MethodImpl(MethodImplOptions.AggressiveInlining)] get; }
     public object? Value { [MethodImpl(MethodImplOptions.AggressiveInlining)] get; }
-    public string Raw { [MethodImpl(MethodImplOptions.AggressiveInlining)] get; }
+    public StringSegment Raw { [MethodImpl(MethodImplOptions.AggressiveInlining)] get; }
     public RegexValue? Regex { [MethodImpl(MethodImplOptions.AggressiveInlining)] get; }
 
-    public string? StringValue { [MethodImpl(MethodImplOptions.AggressiveInlining)] get => TokenType == TokenType.StringLiteral ? (string) Value! : null; }
+    public StringSegment? StringValue { [MethodImpl(MethodImplOptions.AggressiveInlining)] get => TokenType == TokenType.StringLiteral ? (StringSegment) Value! : null; }
     public bool? BooleanValue { [MethodImpl(MethodImplOptions.AggressiveInlining)] get => TokenType == TokenType.BooleanLiteral ? ReferenceEquals(Value, s_boxedTrue) : null; }
     public double? NumericValue { [MethodImpl(MethodImplOptions.AggressiveInlining)] get => TokenType == TokenType.NumericLiteral ? (double) Value! : null; }
     public Regex? RegexValue { [MethodImpl(MethodImplOptions.AggressiveInlining)] get => TokenType == TokenType.RegularExpression ? (Regex?) Value : null; }