--- a/src/Esprima/Ast/PrivateIdentifier.cs
+++ b/src/Esprima/Ast/PrivateIdentifier.cs
@@ -1,16 +1,17 @@
 ﻿using System.Runtime.CompilerServices;
 using Esprima.Utils;
+using Microsoft.Extensions.Primitives;
 
 namespace Esprima.Ast;
 
 public sealed class PrivateIdentifier : Expression
 {
-    public PrivateIdentifier(string name) : base(Nodes.PrivateIdentifier)
+    public PrivateIdentifier(StringSegment name) : base(Nodes.PrivateIdentifier)
     {
         Name = name;
     }
 
-    public string Name { [MethodImpl(MethodImplOptions.AggressiveInlining)] get; }
+    public StringSegment Name { [MethodImpl(MethodImplOptions.AggressiveInlining)] get; }
 
     internal override Node? NextChildNode(ref ChildNodes.Enumerator enumerator) => null;