--- a/src/Esprima/Ast/SyntaxComment.cs
+++ b/src/Esprima/Ast/SyntaxComment.cs
@@ -1,11 +1,12 @@
 ﻿using System.Runtime.CompilerServices;
 using Esprima.Utils;
+using Microsoft.Extensions.Primitives;
 
 namespace Esprima.Ast;
 
 public sealed class SyntaxComment : SyntaxElement
 {
-    public SyntaxComment(CommentType type, string value)
+    public SyntaxComment(CommentType type, StringSegment value)
     {
         Type = type;
         Value = value;
@@ -13,7 +14,7 @@ public sealed class SyntaxComment : SyntaxElement
 
     public CommentType Type { [MethodImpl(MethodImplOptions.AggressiveInlining)] get; }
 
-    public string Value { [MethodImpl(MethodImplOptions.AggressiveInlining)] get; }
+    public StringSegment Value { [MethodImpl(MethodImplOptions.AggressiveInlining)] get; }
 
     public override string ToString()
     {
@@ -26,7 +27,7 @@ public sealed class SyntaxComment : SyntaxElement
                     writer.WriteLineComment(Value, JavaScriptTextWriter.TriviaFlags.None);
                     break;
                 case CommentType.Block:
-                    writer.WriteBlockComment(Value.Split(new[] { "\r\n", "\r", "\n" }, StringSplitOptions.None), JavaScriptTextWriter.TriviaFlags.None);
+                    writer.WriteBlockComment(Value.ToString().Split(new[] { "\r\n", "\r", "\n" }, StringSplitOptions.None), JavaScriptTextWriter.TriviaFlags.None);
                     break;
                 default:
                     throw new InvalidOperationException();