--- a/src/Esprima/Comment.cs
+++ b/src/Esprima/Comment.cs
@@ -1,5 +1,4 @@
-﻿using System.Runtime.CompilerServices;
-using System.Runtime.InteropServices;
+﻿using System.Runtime.InteropServices;
 
 namespace Esprima;