… src/Esprima/JavascriptParser.cs
… src/Esprima/ParserExtensions.cs
… src/Esprima/Scanner.cs
… src/Esprima/Token.cs
… src/Esprima/Utils/JavascriptTextFormatter.cs
… src/Esprima/Utils/JavascriptTextWriter.cs
… src/Esprima/Utils/JsonTextWriter.cs
… src/Esprima/Utils/JsonWriter.cs
… src/Esprima/Utils/KnRJavascriptTextFormatter.cs
… test/Esprima.Tests/AstTests.cs
… test/Esprima.Tests/AstToJavascriptTests.cs
… test/Esprima.Tests/Fixtures.cs
… test/Esprima.Tests/ParserTests.cs