A POC that uses both spans and StringSegment (NuGet package reference still).

* tests still partially broken 
* AST has StringSegment type
* Token has a new reference to string as it already has start and end, can now return both spans and segments