@lahma 2022-08-17
Take these results with some grain of salt as tests are not green. But should show some direction.

## Esprima.Benchmark.FileParsingBenchmark

| **Diff**|Method|FileName|Mean|Error|Gen 0|Gen 1|Allocated|
|------- |-------|-------|-------:|-------|-------:|-------:|-------:|
| Old |ParseProgram|angular-1.2.5|10.428 ms|0.0471 ms|343.7500|156.2500|5,744 KB|
| **New** |	|	| **11.484 ms (+10%)** | **0.0502 ms** | **265.6250 (-23%)** | **109.3750 (-30%)** | **4,395 KB (-23%)** |
| Old |ParseProgram|backbone-1.1.0|1.380 ms|0.0061 ms|52.7344|21.4844|888 KB|
| **New** |	|	| **1.664 ms (+21%)** | **0.0130 ms** | **41.0156 (-22%)** | **19.5313 (-9%)** | **682 KB (-23%)** |
| Old |ParseProgram|jquery-1.9.1|8.243 ms|0.0368 ms|296.8750|140.6250|4,991 KB|
| **New** |	|	| **9.159 ms (+11%)** | **0.0275 ms** | **234.3750 (-21%)** | **93.7500 (-33%)** | **3,877 KB (-22%)** |
| Old |ParseProgram|jquery.mobile-1.4.2|13.418 ms|0.0208 ms|468.7500|234.3750|7,677 KB|
| **New** |	|	| **14.133 ms (+5%)** | **0.0444 ms** | **359.3750 (-23%)** | **171.8750 (-27%)** | **5,940 KB (-23%)** |
| Old |ParseProgram|mootools-1.4.5|6.455 ms|0.0097 ms|242.1875|117.1875|4,028 KB|
| **New** |	|	| **7.287 ms (+13%)** | **0.0149 ms** | **187.5000 (-23%)** | **85.9375 (-27%)** | **3,119 KB (-23%)** |
| Old |ParseProgram|underscore-1.5.2|1.194 ms|0.0021 ms|44.9219|19.5313|748 KB|
| **New** |	|	| **1.388 ms (+16%)** | **0.0039 ms** | **35.1563 (-22%)** | **13.6719 (-30%)** | **578 KB (-23%)** |
| Old |ParseProgram|yui-3.12.0|6.136 ms|0.0257 ms|226.5625|85.9375|3,789 KB|
| **New** |	|	| **6.675 ms (+9%)** | **0.0181 ms** | **179.6875 (-21%)** | **70.3125 (-18%)** | **3,031 KB (-20%)** |