@lahma 2022-08-18
Like discussed, we probably don't want to hold on to potentially large string containing the entire script. String interning and pooling is the chosen approach.