@sebastienros 2022-08-18
Note that parlot falls in this category at the moment. It makes me think that some allocations benchmarks are flawed since we don't take the preallocated memory into account when measuring an iteration.

I haven't looked at the lifetime scope of the pool either in the intern pr. It matters in concurrent systems where the same pool might be in locking contention.